chore: remove duplicated code
